refactor: Cache two render methods directly
